pyproject.toml
[project]
name = "hippocraticum-ai-consultation-sdk"
version = "1.0.0"
description = "A client library for accessing Hippocraticum AI Consultation API"
authors = []
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "httpx>=0.23.0,<0.29.0",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0",
]

[tool.pdm]
distribution = true

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
